+++ front/src/components/Auth/LoginComponents.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const LoginComponents = () => {
+    return <div>Login</div>;
+};
+
+export default LoginComponents;
