+++ front/src/pages/index.jsx
@@ -1,4 +1,4 @@
-export { default as Home } from './Home';
-export { default as Map } from './Map';
-export { default as Login } from './Login';
-export { default as SignUp } from './SignUp';
+export { default as Home } from './Home/Home';
+export { default as Map } from './Map/Map';
+export { default as Login } from './Auth/Login';
+export { default as SignUp } from './Auth/SignUp';
